+++ semcore/d3-chart/src/component/Chart/LineChart.type.ts
@@ -23,6 +23,8 @@ export type LineChartProps = BaseChartProps<LineChartData> & {
   showDots?: boolean;
   curve?: CurveFactory;
   areaCurve?: CurveFactory;
+  /** Callback triggered when a user clicks on a line */
+  onClickLine: (index: number, event: React.SyntheticEvent) => void;
 };
 
 export type LineChartType = Intergalactic.Component<typeof Flex, LineChartProps>;