+++ semcore/d3-chart/src/types/Area.d.ts
@@ -26,6 +26,8 @@ export type AreaProps = Context & {
   transparent?: boolean;
   /** Enables charts patterns that enhances charts accessibility */
   patterns?: PatternsConfig;
+  /** Callback triggered when a user clicks on a chart at a position corresponding to a data item */
+  onClick: (index: number, event: React.SyntheticEvent) => void;
 };
 
 /** @deprecated */