+++ semcore/d3-chart/src/types/Donut.d.ts
@@ -28,6 +28,8 @@ export type DonutProps = Context & {
   paddingAngle?: number;
   /** Enables charts patterns that enhances charts accessibility */
   patterns?: PatternsConfig;
+  /** Callback triggered when a user clicks on a pie */
+  onClick: (key: string, event: React.SyntheticEvent) => void;
 };
 
 /** @deprecated */