+++ semcore/d3-chart/src/types/Line.d.ts
@@ -27,6 +27,8 @@ export type LineProps = Context & {
   transparent?: boolean;
   /** Enables charts patterns that enhances charts accessibility */
   patterns?: PatternsConfig;
+  /** Callback triggered when a user clicks on a line */
+  onClick: (index: number, event: React.SyntheticEvent) => void;
 };
 
 /** @deprecated */