+++ semcore/d3-chart/src/types/Radar.d.ts
@@ -39,6 +39,8 @@ export type RadarProps = Context & {
   angleOffset?: number;
   /** Enables charts patterns that enhances charts accessibility */
   patterns?: PatternsConfig;
+  /** Callback triggered when a user clicks on a chart at a position corresponding to a category */
+  onClick: (index: number, e: React.SyntheticEvent) => void;
 };
 
 /** @deprecated */